Move recipe suggetions to seperate file
